requirements.txt
alpaca-py==0.7.0
amqp==5.1.1
APScheduler==3.9.1.post1
async-timeout==4.0.2
billiard==3.6.4.0
boto3==1.26.56
botocore==1.29.56
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contourpy==1.0.7
cycler==0.11.0
exchange-calendars==4.2.5
Flask==2.2.2
Flask-HTTPAuth==4.7.0
fonttools==4.38.0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
kombu==5.2.4
korean-lunar-calendar==0.3.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
msgpack==1.0.4
mysql-connector==2.2.9
numpy==1.24.1
packaging==23.0
pandas==1.5.3
pandas-market-calendars==4.1.4
Pillow==9.4.0
prompt-toolkit==3.0.38
pycparser==2.21
pydantic==1.10.4
pyluach==2.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
redis==4.5.1
requests==2.28.2
s3transfer==0.6.0
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
sklearn==0.0.post1
sseclient-py==1.7.2
supervisor==4.2.5
threadpoolctl==3.1.0
toolz==0.12.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
websockets==10.4
Werkzeug==2.2.2
